Migrate from UDP to TCP (#11)
* Rewrite networking into TCP

* Use the whole rollback window

The game is now playable

* Refactors

* Update AppStateIngame.cpp

* Fix server-test

* Fix crash when remote peer disconnects

* Fix join game dialog not working after the initial attempt was unsuccesfull

There was a bug in dgm-lib

+++ src/lib-network/include/LobbySettings.hpp
@@ -20,7 +20,7 @@ struct [[nodiscard]] LobbySettings final
     std::vector<MapSettings> mapSettings = {};
     std::vector<size_t> mapOrder = {};
     int fraglimit = 0;
-    unsigned maxNpcs = 0;
+    size_t maxNpcs = 0;
 
     [[nodiscard]] constexpr std::partial_ordering
     operator<=>(const LobbySettings&) const = default;
