+++ src/lib-options/include/CmdParameters.hpp
@@ -11,6 +11,6 @@ struct CmdParameters
     std::filesystem::path demoFile = "demo.txt";
     bool playDemo = false;
     bool enableDebug = false;
-    unsigned maxNpcs = 3;
+    size_t maxNpcs = 3;
     unsigned fraglimit = 15;
 };
